Partial implementation of real time improvements to speech trimmer.
Basically, I'm trying to make it so you can stuff any size of real time data
into speech trimmer rather than just a centisecond worth of data at a time.
It's not that the algorithm is to slow or not responsive.  It isn't.  This just
makes it less work to use.